requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-google-genai
google-api-python-client
google-auth
pydantic
python-dotenv
httpx
pydantic-settings